feat(api): implement cursor-based pagination
We ignore duplicated creation dates - it's unlikely two wishes are
created at exactly the same timestamp and even if they are - it's not
really that crucial to show them all. It'd be difficult to implement it
with the current architecture when the wish ID is a string without
any particular ordering.